+++ stream-chat-android-ui-components-sample/src/main/kotlin/io/getstream/chat/ui/sample/util/extensions/Channel.kt
@@ -18,4 +18,4 @@ package io.getstream.chat.ui.sample.util.extensions
 
 import io.getstream.chat.android.offline.model.channel.ChannelData
 
-public fun ChannelData.isAnonymousChannel(): Boolean = channelId.startsWith("!members")
+public fun ChannelData.isAnonymousChannel(): Boolean = id.startsWith("!members")