+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/channel/actions/internal/ChannelActionsViewModelFactory.kt
@@ -23,11 +23,9 @@ import androidx.lifecycle.ViewModelProvider
  * Specialized factory class that produces [ChannelActionsViewModel].
  *
  * @param cid The full channel id, i.e. "messaging:123".
- * @param isGroup True if the Channel is a group channel, false otherwise.
  */
 internal class ChannelActionsViewModelFactory(
     private val cid: String,
-    private val isGroup: Boolean,
 ) : ViewModelProvider.Factory {
 
     /**
@@ -35,6 +33,6 @@ internal class ChannelActionsViewModelFactory(
      */
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return ChannelActionsViewModel(cid, isGroup) as T
+        return ChannelActionsViewModel(cid) as T
     }
 }